Fix copy construction of the Nullable and Stringifyable
We have to split cunstruction: from the same type and completely other
type to prevent loosing the value of the private flags.

This issue if VS2010 specific. On e.g. VS2013 this constructor is not
called at all. Possibly some kind of copy elision take place.

The methods were disabled with help of "default arg" approach
since default templete method param doesn't work in VS2010(it works only
for the class templates).

Method implementations were moved to the declaration because VS2010 fails
to compile the code when the definition and declaration are split.
Error: error: C2244: 'rpc::Nullable<T>::{ctor}' : unable to match function
definition to an existing declaration.